Sample Tescase executed.
Sample Tescase executed.

+++ service/src/main/java/au/com/project/sample/persistence/PersonDAO.java
@@ -26,6 +26,6 @@ public interface PersonDAO {
 
     List<Person> findRangePerson(int[] range);
 
-    int countPerson();
+    Long countPerson();
 
 }
